requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.4
django-ckeditor==6.7.0
django-heroku==0.3.1
django-js-asset==2.1.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.7
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
